Add order-item deduplication endpoint and batch deletion support

+++ craftstudio-mes-application/src/main/java/com/mes/application/command/order/vo/OrderItemDeduplicationResult.java
@@ -11,6 +11,6 @@ public class OrderItemDeduplicationResult {
     private String orderId;
     private String keptOrderItemId;
     private int matchedOrderItemCount;
-    private int deletedOrderItemCount;
+    private long deletedOrderItemCount;
     private long deletedProductionPieceCount;
 }
